refactor: split design proposal contract validators
